+++ src/models/courses/actions/index.js
@@ -1,0 +1,3 @@
+export { getCourses, loadingCoursesStarted, loadingCoursesSucceed, loadingCoursesFailed } from './actions-get';
+export { createCourse, creatingCourseStarted, creatingCourseSucceed, creatingCourseFailed } from './actions-post';
+export { editCourse, editingCourseStarted, editingCourseSucceed, editingCourseFailed } from './actions-put';
