+++ src/models/courses/index.js
@@ -1,4 +1,4 @@
 export { coursesReducer } from './reducer.js';
-export { getCourses, createCourse, editCourse } from './actions.js';
 export { loadCoursesWatcher, createCourseWatcher, editCourseWatcher } from './sagas.js';
 export { coursesDataSelector, coursesIsLoadingSelector } from './selectors.js';
+export { getCourses, createCourse, editCourse} from './actions/index.js';
